+++ components/projects/form/fields/MemberField.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { isEmpty } from 'lodash-es';
 import { FC, useMemo, useState } from 'react';
 
-import { getMembersSearchByName } from '@/api/endpoint/members';
+import { getMembersSearchByName } from '@/api/endpoint/members/getMembersSearchByName';
 import Input from '@/components/common/Input';
 import ErrorMessage from '@/components/common/Input/ErrorMessage';
 import Select from '@/components/common/Select';
