+++ components/projects/upload/hooks/useGetMembersByNameQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getMembersSearchByName } from '@/api/endpoint/members';
+import { getMembersSearchByName } from '@/api/endpoint/members/getMembersSearchByName';
 interface GetMembersByNameQueryVariables {
   name: string;
 }
